requirements.txt
opencv-python>=4.10.0
Pillow>=11.0.0
numpy>=2.1.1
